Extract error tests to test/unit/core
